+++ samples/android-simple/Silencer/app/src/main/java/com/muflihun/silencer/ConnectActivity.java
@@ -22,7 +22,7 @@ public class ConnectActivity extends AppCompatActivity {
         System.out.println("reconnect()...");
 
         r.setAccessCodeMap(new HashMap<String, String>() {{
-            put("sample-app", "eif89");
+            put("sample-app", "a2dcb");
         }});
 
         r.setHost(hostText.getText().toString(), Integer.valueOf(portText.getText().toString()));